fix test coverage
